# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dental Clinic Scheduling API",
        "description": "API REST para gestionar citas de una clinica dental. Permite autenticacion, creacion de pacientes, creacion de citas y listado de citas por dia.",
        "contact": {
            "email": "recepcionista@pruebasmulhacen.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local"
        }
    ],
    "paths": {
        "/api/appointments": {
            "get": {
                "tags": [
                    "Citas"
                ],
                "summary": "Listar citas por dia",
                "description": "Devuelve todas las citas de un dia concreto, ordenadas por hora de inicio. Requiere autenticacion.",
                "operationId": "812c4f15927f3d0e3e959538a6f1104b",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Fecha en formato Y-m-d",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-03-15"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de citas del dia",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Citas del día 2026-03-15."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "example": "2026-03-15 09:00"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "example": "2026-03-15 09:45"
                                                    },
                                                    "dentist_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "patient_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "treatment_ids": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "integer"
                                                        },
                                                        "example": [
                                                            1
                                                        ]
                                                    },
                                                    "reason": {
                                                        "type": "string",
                                                        "example": "Colocacion de brackets"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Parametro date requerido o formato invalido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Citas"
                ],
                "summary": "Crear cita",
                "description": "Crea una nueva cita dental. Valida solapamiento de horarios por dentista. Requiere autenticacion.",
                "operationId": "2b481df28ec53d2afb2ec7c9a3a77846",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id",
                                    "dentist_id",
                                    "start_time",
                                    "duration",
                                    "reason",
                                    "treatment_ids"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "dentist_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "description": "Formato Y-m-d H:i",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-03-15 09:00"
                                    },
                                    "duration": {
                                        "description": "Duracion en minutos",
                                        "type": "integer",
                                        "maximum": 480,
                                        "minimum": 5,
                                        "example": 45
                                    },
                                    "reason": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "Colocacion de brackets"
                                    },
                                    "treatment_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "minItems": 1,
                                        "example": [
                                            1
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cita creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cita creada correctamente."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "start_time": {
                                                    "type": "string",
                                                    "example": "2026-03-15 09:00"
                                                },
                                                "end_time": {
                                                    "type": "string",
                                                    "example": "2026-03-15 09:45"
                                                },
                                                "dentist_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "patient_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "treatment_ids": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer"
                                                    },
                                                    "example": [
                                                        1
                                                    ]
                                                },
                                                "reason": {
                                                    "type": "string",
                                                    "example": "Colocacion de brackets"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "409": {
                        "description": "Solapamiento de horario con otra cita del mismo dentista",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El horario seleccionado no está disponible para este dentista."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validacion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtenido via POST /api/login",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}